For adoption data from Excel files, 0.0 == NaN.
In the Excel implementation, adoption data of 0.0 is treated
the same as N/A, no data available.

We don't want to implement adoptiondata.py the same way, we
want to be able to express the difference between a solution
which did not exist prior to year N, and therefore had 0.0
adoption, from a solution which did exist but for which we
have no data prior to year N.

We're handling this in the code generator: when extracting
adoption data from an Excel file, treat values of 0.0 as N/A
and write out a CSV file with no data at that location. The
Python implementation will then process N/A for those years,
the same as Excel uses.