jupyter: toplevel overview & replace graphviz.
Move the overview (the list of all solutions and the
donut chart + treemap) out of the tab bar and onto the
main page. This is a step toward using the solution list
as the way to choose solutions to examine in more detail,
removing the selection box of solution+scenarios.

Also:
+ replace the graphviz SVG.
  The graphviz model diagram worked, but did not spark joy.
  Replace it with a nicer one generated using Omnigraffle.
+ move giant block of solution imports into a factory.
  solution/factory.py can return a list of all solutions,
  with sufficient information to construct an object for any
  of them.
+ split Vega code (treemap and nested donut chart) into its
  own file and tests.