customadoption: implement total adoption limit.
Custom Adoption needs to impose an upper limit on adoption, which
will typically be set from the TAM for RRS solutions and from the
TLA for Land solutions. Implement a total_adoption_limit argument,
with a unit test.

Noticed while working on Tree Intercropping, which hits the TLA
in its third scenario.

Remove comment about regional TLA, as a similar comment appears
in a number of places in the codebase now. The comment in tla.py is
the relevant one, and the one which is likely to be removed when
some sort of regional support is implemented. The one in customadoption
is not applicable to many of the solutions which use custom adoption,
and likely to be overlooked when a regional TLA is worked out so then
we have a stale but scary comment.