All solution class names are Scenario.
A huge portion of the time taken to start the UI and render
the overview of solutions+sectors was spent importing the solution
modules in solution/factory.py, in order to get the constructor and
names of the scenarios.

Standardize on a class name of "Scenario" for all solutions, allowing
solution/factory.py to rely on data/overview/solutions.csv and not
always have to import every solution.

UI startup is considerably faster.