FaIR: remove FaIR module from all solutions.
This reverts commit 2fc0eda15dee0446e9a82fccb3120ced671469ce.

We initially wrote FaIR as a seperate model object because we
expected to set useMultigas=True and pass in CO2, CH4, N2O, HFC,
etc. We're not going to be able to do that right away as we don't
have all of the data which would be required, so we're currently
using useMultigas=False and passing in CO2eq.

As such, remove the separate FaIR object from solutions. We're going
to incorporate the functionality into model/co2calcs.py until it
gets to the point where a separate object is called for.