tam: better regional support
Remove knowledge of the names of the regions from model/tam.py.
TAM will now work with whatever model/dd.py contains.

We initially included explicit routines for each region because
we structured the Python code as a Flask webserver which Excel
did an HTTP fetch from. We named the functions the same as the
JSON variables they produced. The Flask webserver is long gone,
we now call into the Python code directly and no longer use Excel.